Implemented Shop Upgrade Purchase
EnemyMovement:
~ changed "health" to float

~ changed "TakeDamage()" to recieve float arguments

PlayerController:
+ int "wallet"

+ float "damageModifier"

+ float "baseDamage"

+ SpendCurrency(int amount)

~ "EquipItem()" currently directly modifies "damageModifier" for temp progress simulation

PlayerHitDetection:
~ changed call to "em.TakeDamage(25)" to "em.TakeDamage(pc.baseDamage * pc.attackModifier)"

UI_Shop:
+ added "BuyItem(int cost)"

+ added reference to "pc" to call "spendCurrency()" method from PlayerController.cs

+++ term-project-main/Assets/Scripts/EnemyMovement.cs
@@ -9,7 +9,7 @@ public class EnemyMovement : MonoBehaviour
 {
     [Header("Stats")]
     public Bartender enemyHealth;
-    [Range(0, 100)] public int health = 100;
+    [Range(0, 100)] public float health = 100;
 
     [Header("Nav Agent Vars")]
     [Range(2, 5)] public float agentSpeed;
@@ -61,7 +61,7 @@ public class EnemyMovement : MonoBehaviour
     /* ---------------------------------------------------------------- */
     /*                           Helper Methods                         */
     /* ---------------------------------------------------------------- */
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         health -= amount;
         enemyHealth.setValue(health);
